# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Chef~Gaustau's Restaurant API",
        "description": "An online food delivery app for a restaurant",
        "license": {
            "name": "MIT",
            "url": "https://github.com/ArnauldAnguh/Restau-apis/blob/develop/LICENSE"
        }
    },
...
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "schemes": ["http"],
    "consumes": [
        "application/json",
        "application/x-www-form-urlencoded"
    ],
    "produces": [
        "application/json",
        "application/xml",
        "text/html"
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": {
                    "id": "id",
                    "name": "name",
                    "email": "email",
                    "password": "password"
                }
            },
            "Fooditems": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "qunatity": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "integer"
                    }
                },
                "required": {
                    "name": "name",
                    "image": "image",
                    "description": "description",
                    "price": "price",
                    "quantity": "quantity"
                }
            },
            "Orders": {
                "type": "object",
                "properties": {
                    "quantity": {
                        "type": "integer"
                    },
                    "unit_price": {
                        "type": "integer"
                    },
                    "fooditem": {
                        "type": "string"
                    }
                },
                "required": {
                    "unit_price": "unit_price",
                    "quantity": "quantity",
                    "fooditem": "fooditem"
                }
            }
        }
    },
...
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "operationId": "getUsers",
                "summary": "Get all customer users in the system",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                    },
                    "405": {
                        "type": "string",
                        "description": "No users yet"
                    },
                    "500": {
                        "type": "string",
                        "description": "InternalServerError"
                    }
                }
            }
        },
...
        "/foodItems": {
            "get": {
                "tags": [
                    "Food Items"
                ],
                "operationId": "getAllItems",
                "summary": "Get all food items in the system",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/components/schemas/Fooditems"
                        }
                    },
                    "405": {
                        "type": "string",
                        "description": "No items yet"
                    },
                    "500": {
                        "type": "string",
                        "description": "InternalServerError"
                    }
                }
            }
        },
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "operationId": "fetchAllOrders",
                "summary": "Get all orders in the system",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/components/schemas/Orders"
                        }
                    },
                    "405": {
                        "type": "string",
                        "description": "No Orders yet"
                    },
                    "500": {
                        "type": "string",
                        "description": "InternalServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "operationId": "placeOrder",
                "summary": "place a new Order",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "place a new Order in the system",
                    "required": true,
                    "schema": {
                        "$ref": "#/components/schemas/Orders"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "schema": {
                            "$ref": "#/components/schemas/Orders"
                        }
                    },
                    "UnauthorizedError": {
                        "description": "Access token is missing or invalid"
                    },
                    "500": {
                        "description": "internalServerError"
                    }
                }
            }
        },
...
    }
}